Add MFEMWeakForm base class
